+++ magik-squid/src/test/java/nl/ramsolutions/sw/magik/analysis/typing/reasoner/LocalTypeReasonerTest.java
@@ -1,7 +1,6 @@
 package nl.ramsolutions.sw.magik.analysis.typing.reasoner;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.Assert.assertThat;
 
 import com.sonar.sslr.api.AstNode;
 import java.io.IOException;
@@ -16,9 +15,7 @@ import nl.ramsolutions.sw.magik.analysis.definitions.ExemplarDefinition;
 import nl.ramsolutions.sw.magik.analysis.definitions.IDefinitionKeeper;
 import nl.ramsolutions.sw.magik.analysis.definitions.MethodDefinition;
 import nl.ramsolutions.sw.magik.analysis.definitions.ParameterDefinition;
-import nl.ramsolutions.sw.magik.analysis.definitions.ProcedureDefinition;
 import nl.ramsolutions.sw.magik.analysis.definitions.SlotDefinition;
-import nl.ramsolutions.sw.magik.analysis.definitions.TypeStringDefinition;
 import nl.ramsolutions.sw.magik.analysis.typing.ExpressionResultString;
 import nl.ramsolutions.sw.magik.analysis.typing.TypeString;
 import nl.ramsolutions.sw.magik.api.MagikGrammar;
@@ -659,14 +656,7 @@ class LocalTypeReasonerTest {
     assertThat(result)
         .isEqualTo(
             new ExpressionResultString(
-                TypeString.ofIdentifier(TypeString.ANONYMOUS_PACKAGE, "_proc_in_memory_0")));
-
-    final TypeString procTypeStr = result.get(0, null);
-    final TypeStringDefinition typeStringDefinition = state.getTypeStringDefinition(procTypeStr);
-    assertThat(typeStringDefinition).isExactlyInstanceOf(ProcedureDefinition.class);
-    final ProcedureDefinition procDef = (ProcedureDefinition) typeStringDefinition;
-    final ExpressionResultString procResult = procDef.getReturnTypes();
-    assertThat(procResult).isEqualTo(new ExpressionResultString(TypeString.SW_INTEGER));
+                TypeString.ofIdentifier("_proc_in_memory_0", TypeString.ANONYMOUS_PACKAGE)));
   }
 
   @Test