+++ CHANGES.md
@@ -23,9 +23,11 @@
   - Hovering.
   - Definition.
   - References.
+  - More robust grammar.
 - Remove old typing classes and use the new typing classes.
 - Fix providing Formatting CodeActions on all locations, instead of asked range.
 - Fix hovering over atoms.
+- Several fixes.
 
 0.9.1 (2024-03-13)